requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
apturl==0.5.2
arrow==1.2.3
asyncpg==0.26.0
atlastk==0.13.2
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
binaryornot==0.4.4
blinker==1.4
boto3==1.24.60
botocore==1.27.60
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cookiecutter==2.1.1
cryptography==2.8
cssselect==1.2.0
cssutils==2.6.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
entrypoints==0.3
evdev==1.3.0
fastapi==0.79.1
fastapi-jwt-auth==0.5.0
fasteners==0.14.1
fireo==1.5.1
folder-color-common==0.0.88
future==0.18.2
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.3.4
googleapis-common-protos==1.56.4
greenlet==2.0.1
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.9.0
html5lib==1.0.1
httplib2==0.14.0
httptools==0.1.2
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
jose==1.0.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.1.2
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==1.0.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
Pint==0.20.1
premailer==3.10.0
proto-plus==1.22.0
protobuf==4.21.5
psutil==5.5.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pypika-tortoise==0.1.6
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.0.0-ubuntu0.20.04.7-zorin1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-jose==3.3.0
python-magic==0.4.16
python-multipart==0.0.5
python-resize-image==1.1.20
python-slugify==8.0.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
rsa==4.9
s3transfer==0.6.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
shortuuid==1.0.11
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==1.9.5
speedtest-cli==2.1.2
SQLAlchemy==1.4.41
sqlalchemy-orm==1.2.5
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
ssh-import-id==5.10
starlette==0.19.1
stravalib==1.2.0
text-unidecode==1.3
texttable==1.6.2
tortoise==0.1.1
tortoise-orm==0.19.2
typing-extensions==4.3.0
typing-inspect==0.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.16.0
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
websockets==8.1
wsproto==0.15.0
xkit==0.0.0
zipp==1.0.0
zorin-appearance==4.1
zorin-connect==1.2
zorin-exec-guard==1.4
